+++ tests/forcing/test_pcrglobwb.py
@@ -36,10 +36,11 @@ def mock_recipe_run(monkeypatch, tmp_path):
             create_netcdf("tas", tmp_path / "pcrglobwb_tas.nc"),
         )
 
-    def mock_run(self):
+    def mock_run(self, session=None):
         """Store recipe for inspection and return dummy output."""
         nonlocal data
         data["data_during_run"] = self.data
+        data["session"] = session
         return {"diagnostic_daily/script": MockTaskOutput()}
 
     monkeypatch.setattr(Recipe, "run", mock_run)
@@ -81,3 +82,27 @@ class TestGenerateWithExtractRegion:
 
 # TODO test if recipe was generated correctlu
 # TODO test if yaml was written
+
+
+def test_with_directory(mock_recipe_run, sample_shape, tmp_path):
+    forcing_dir = tmp_path / "myforcing"
+    generate(
+        target_model="pcrglobwb",
+        dataset="ERA5",
+        start_time="1989-01-02T00:00:00Z",
+        end_time="1999-01-02T00:00:00Z",
+        shape=str(sample_shape),
+        directory=forcing_dir,
+        model_specific_options=dict(
+            start_time_climatology="1979-01-02T00:00:00Z",
+            end_time_climatology="1989-01-02T00:00:00Z",
+            extract_region={
+                "start_longitude": 10,
+                "end_longitude": 16.75,
+                "start_latitude": 7.25,
+                "end_latitude": 2.5,
+            },
+        ),
+    )
+
+    assert mock_recipe_run["session"].session_dir == forcing_dir